+++ Documentation/Changelog.md
@@ -33,6 +33,7 @@ Release date: not released yet
 - .NET Framework 4.6 support is retired. The minimal supported .NET Framework version is 4.6.1
 - Fixed case-sensitivity of a text data encoding padding character. This is important for data encodings that use custom padding characters
 - Fixed issue in `Gapotchenko.FX.Collections.Generic.AssociativeArray<TKey, TValue>` type with accessing `IEnumerable<T>.Current` property without checking the result of a prior call to `MoveNext` method
+- Fixed assembly name comparison bug in `Gapotchenko.FX.Reflection.Loader` module
 
 ## 2022